concesionario/concesionario/requirements.txt
Django==1.8.4
django-appconf==1.0.1
django-imagekit==3.2.7
pilkit==1.1.13
Pillow==2.9.0
six==1.9.0
wheel==0.24.0
